extract optInDeletionChoiceHandler saga
